refactor/cleaner owned by & policy

+++ app/Http/Controllers/PostController.php
@@ -38,9 +38,7 @@ class PostController extends Controller
 
 	public function delete(Post $post)
 	{
-		if ($post->ownedBy(auth()->user())) {
-			dd('no');
-		}
+		$this->authorize('delete', $post);
 		$post->delete();
 		return back();
 	}
